Very Close to completion
Formatted all of the backgrounds by adding an image, color, or gradients. Formatted the pages with grids and and floating items. Added all of the content into the webpage including images. Added in filters to all of the images to make them pop more. Added transforms to some of the images to make them look  better. Added shadows around the body and around the image text. Added a border around the address box.